Set 'nodata' for NA values in export image - provided R script for 0 to NA conversion not required anymore!
In previous version (time of publication), the exported images contained 0 (zero) where there was actually a "no data" value. Therefore, we provided an R script to convert 0 values to NA.

Now we've updated the code, so that images that are exported to GoogleDrive already contain NA values where there is no valid pixel available. Hence there is no need anymore to use the provided R script to convert 0 to NA.